Fixes #5888: Only compile in the dev mode logic if you build with “-—tags dev”
This lets you have quicker cycles on if your testing from your desktop.  But if if it’s compiled into prod mode, it panics since it can’t find the source files.
