+++ install/operator/pkg/generator/mode_prod.go
@@ -1,0 +1,11 @@
+// +build !dev
+
+package generator
+
+import (
+	"net/http"
+)
+
+func GetAssetsFS() http.FileSystem {
+	return assets
+}
